Update Active Directory User attributes in a bulk modus. Data is provide via a CSV - file with result output on screen or logging.
The Import-ADAttributes will update Active Directory User attributes in
a bulk modus. The data has been provide by a CSV file.
The headers of the CSV file will be the AD User attributes. The import
process will import the data depending on the headers in the file.
So no pre - defined fields needed in the script, just put them in the
header of the CSV - file.  The result can been shown on your screen or
can be stored in a logging file.